Move circuit CLI's REST API client up
Moves the `SplinterRestClient` that was previously a private struct of
the circuit CLI up to its own module in the `action` module. This makes
the client available to other commands that need query a splinter node.

The new client includes a `get_node_status` method that replaces the
`fetch_node_id` method for querying the splinter node's `/status`
endpoint.
