+++ cli/src/action/mod.rs
@@ -12,7 +12,9 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
 pub mod admin;
+mod api;
 pub mod certs;
 pub mod circuit;
 #[cfg(feature = "database")]
